+++ packages/typir/src/graph/graph-algorithms.ts
@@ -13,10 +13,37 @@ import { Type } from './type-node.js';
 /**
  * Graph algorithms to do calculations on the type graph.
  * All algorithms are robust regarding cycles.
+ * Only edges marked with `LINK_EXISTS` are considered by default.
  */
 export interface GraphAlgorithms {
+    /**
+     * Collects all direct and transitively reachable types for a given type.
+     * @param from the starting point
+     * @param $relations the kinds of edges to traverse
+     * @param filterEdges optionally ignore some edges
+     * @returns If there are cycles, the result will include the given starting point, otherwise this type is not included.
+     */
     collectReachableTypes(from: Type, $relations: RelationInformation[], filterEdges?: (edgr: TypeEdge) => boolean): Set<Type>;
+
+    /**
+     * Searches whether there is a non-empty path from a given start type to the given end type.
+     * @param from the starting node type
+     * @param to the end/destination node type
+     * @param $relations the kinds of edges to traverse
+     * @param filterEdges optionally ignore some edges
+     * @returns true, if a non-empty path exists, false otherwise
+     */
     existsEdgePath(from: Type, to: Type, $relations: RelationInformation[], filterEdges?: (edgr: TypeEdge) => boolean): boolean;
+
+    /**
+     * Searches for a non-empty path from the given start type to the given end type.
+     * @param from the starting node type
+     * @param to the end/destination node type
+     * @param $relations the kinds of edges to traverse
+     * @param filterEdges optionally ignore some edges
+     * @returns an empty array indicates, that no path was found,
+     * otherwise the list of found edges in the type graph connecting the starting type with the end type
+     */
     getEdgePath(from: Type, to: Type, $relations: RelationInformation[], filterEdges?: (edgr: TypeEdge) => boolean): TypeEdge[];
 }
 
@@ -122,23 +149,38 @@ export class DefaultGraphAlgorithms<Specifics extends TypirSpecifics> implements
         return [];
     }
 
-    protected calculateRelevantEdges($relation: RelationInformation, current: Type, filterEdges?: (edgr: TypeEdge) => boolean): Array<{ edge: TypeEdge, myEnd: Type, otherEnd: Type }> {
-        if ($relation.direction === 'Unidirectional') {
+    protected calculateRelevantEdges($relation: RelationInformation, current: Type, filterEdges?: (edge: TypeEdge) => boolean): Array<{ edge: TypeEdge, myEnd: Type, otherEnd: Type }> {
+        const check = this.calculateRelevantEdgesCondition($relation, current, filterEdges);
+        if ($relation.direction === 'UnidirectionalFromTo') {
+            // for directed edges: follow edges in the direction "from --> to"
             return current.getOutgoingEdges($relation.$relation)
-                .filter(e => e.cachingInformation === 'LINK_EXISTS' && (filterEdges === undefined || filterEdges(e)))
+                .filter(check)
                 .map(e => ({ edge: e, myEnd: e.from, otherEnd: e.to }));
-        }
-        if ($relation.direction === 'Bidirectional') {
+        } else if ($relation.direction === 'UnidirectionalToFrom') {
+            // for directed edges: follow edges in the direction "to --> from"
+            return current.getIncomingEdges($relation.$relation)
+                .filter(check)
+                .map(e => ({ edge: e, myEnd: e.to, otherEnd: e.from }));
+        } else if ($relation.direction === 'Bidirectional') {
             // for bidirectional edges, both outgoing and incoming edges need to be checked, while to and from are swapped
             return [
                 ...current.getOutgoingEdges($relation.$relation)
-                    .filter(e => e.cachingInformation === 'LINK_EXISTS' && (filterEdges === undefined || filterEdges(e)))
+                    .filter(check)
                     .map(e => ({ edge: e, myEnd: e.from, otherEnd: e.to })),
                 ...current.getIncomingEdges($relation.$relation)
-                    .filter(e => e.cachingInformation === 'LINK_EXISTS' && (filterEdges === undefined || filterEdges(e)))
+                    .filter(check)
                     .map(e => ({ edge: e, myEnd: e.to, otherEnd: e.from })),
             ];
+        } else {
+            assertUnreachable($relation.direction);
         }
-        assertUnreachable($relation.direction);
     }
+
+    protected calculateRelevantEdgesCondition(_$relation: RelationInformation, _start: Type, filterEdges?: (edge: TypeEdge) => boolean): (edge: TypeEdge) => boolean {
+        // by default, only existing edges are considered
+        return filterEdges
+            ? edge => edge.cachingInformation === 'LINK_EXISTS' && filterEdges.call(filterEdges, edge)
+            : edge => edge.cachingInformation === 'LINK_EXISTS';
+    }
+
 }