+++ packages/typir/src/services/conversion.ts
@@ -96,10 +96,10 @@ export interface TypeConversion {
     isConvertible(from: Type, to: Type): boolean;
 
     /**
-     * Returns all other types to which the given type can be recursively converted.
+     * Returns all other types to which the given type can be directly or transitively converted.
      * @param from the source type, which is convertible to the returned types
      * @param mode only conversion rules with the given conversion mode are considered
-     * @returns the set of recursively reachable types for conversion ("conversion targets")
+     * @returns the set of transitively reachable types for conversion ("conversion targets")
      */
     getConvertibleTo(from: Type, mode: ConversionModeForSpecification): Set<Type>;
 }
@@ -178,11 +178,11 @@ export class DefaultTypeConversion<Specifics extends TypirSpecifics> implements 
     }
 
     protected collectReachableTypes(from: Type, mode: ConversionModeForSpecification): Set<Type> {
-        return this.algorithms.collectReachableTypes(from, [{ $relation: ConversionEdge, direction: 'Unidirectional' }], edge => (edge as ConversionEdge).mode === mode);
+        return this.algorithms.collectReachableTypes(from, [{ $relation: ConversionEdge, direction: 'UnidirectionalFromTo' }], edge => (edge as ConversionEdge).mode === mode);
     }
 
     protected existsEdgePath(from: Type, to: Type, mode: ConversionModeForSpecification): boolean {
-        return this.algorithms.existsEdgePath(from, to, [{ $relation: ConversionEdge, direction: 'Unidirectional' }], edge => (edge as ConversionEdge).mode === mode);
+        return this.algorithms.existsEdgePath(from, to, [{ $relation: ConversionEdge, direction: 'UnidirectionalFromTo' }], edge => (edge as ConversionEdge).mode === mode);
     }
 
     protected isTransitivelyConvertable(from: Type, to: Type, mode: ConversionModeForSpecification): boolean {